+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Fixed
+
+* Guards previous focused element.
+
 ## [1.1.0][] - 2018-04-04
 
 ### Added